test(sync): extract SafetensorsFixture so four suites share one byte layout
Safetensors()/Meta() were copy-pasted verbatim from SafetensorsHeaderReaderTests
into IdentifyModelStepTests (the comment admitted it), and the BundledSnapshot
reflection block was duplicated between BaseModelHeaderMapTests and
FilenameBaseModelHeuristicTests. The safetensors byte layout is exactly the
thing that must not drift between suites, and nothing stopped it from doing so.

New SafetensorsFixture (Sync/Service/Identity) holds both builders plus a
CatalogLabels accessor; all four test classes now use it instead of their
own copy.
