+++ DiffusionNexus.Tests/Sync/Service/Identity/BaseModelHeaderMapTests.cs
@@ -1,5 +1,3 @@
-using System.Reflection;
-using DiffusionNexus.Civitai;
 using DiffusionNexus.Service.Services.Sync.Identity;
 using FluentAssertions;
 
@@ -115,19 +113,15 @@ public class BaseModelHeaderMapTests
     }
 
     /// <summary>
-    /// Reflects over <see cref="CivitaiBaseModelCatalog"/>'s private bundled snapshot (rather than
-    /// duplicating the label list here) and asserts every label the map can ever return is a real
-    /// Civitai display label, not a typo'd string nobody's dropdown recognizes.
+    /// Reflects over <see cref="DiffusionNexus.Civitai.CivitaiBaseModelCatalog"/>'s private bundled
+    /// snapshot (via <see cref="SafetensorsFixture.CatalogLabels"/>, rather than duplicating the
+    /// label list here) and asserts every label the map can ever return is a real Civitai display
+    /// label, not a typo'd string nobody's dropdown recognizes.
     /// </summary>
     [Fact]
     public void Map_EveryOutputIsACatalogLabel()
     {
-        var bundledSnapshotField = typeof(CivitaiBaseModelCatalog)
-            .GetField("BundledSnapshot", BindingFlags.NonPublic | BindingFlags.Static);
-        bundledSnapshotField.Should().NotBeNull(
-            "CivitaiBaseModelCatalog.BundledSnapshot must exist for this check to mean anything");
-
-        var bundledSnapshot = (IReadOnlyList<string>)bundledSnapshotField!.GetValue(null)!;
+        var bundledSnapshot = SafetensorsFixture.CatalogLabels;
 
         BaseModelHeaderMap.AllLabels.Should().NotBeEmpty();
         foreach (var label in BaseModelHeaderMap.AllLabels)
